Move sensor env config to a separate file
