+++ util.go
@@ -178,84 +178,3 @@ func hexGatewayToAddr(gateway []rune) (string, error) {
 
 	return fmt.Sprintf("%v.%v.%v.%v", octets[0], octets[1], octets[2], octets[3]), nil
 }
-
-// parseInstanaTags parses the tags string passed via INSTANA_TAGS.
-// The tag string is a comma-separated list of keys optionally followed by an '=' character and a string value:
-//
-//     INSTANA_TAGS := key1[=value1][,key2[=value2],...]
-//
-// The leading and trailing space is truncated from key names, values are used as-is. If a key does not have
-// value associated, it's considered to be nil.
-func parseInstanaTags(tagStr string) map[string]interface{} {
-	tags := make(map[string]interface{})
-
-	for _, s := range strings.Split(tagStr, ",") {
-		kv := strings.SplitN(s, "=", 2)
-
-		k := strings.TrimSpace(kv[0])
-		if k == "" {
-			continue
-		}
-
-		var v interface{}
-		if len(kv) > 1 {
-			v = kv[1]
-		}
-
-		tags[k] = v
-	}
-
-	if len(tags) == 0 {
-		return nil
-	}
-
-	return tags
-}
-
-// parseInstanaSecrets parses the tags string passed via INSTANA_SECRETS.
-// The secrets matcher configuration string is expected to have the following format:
-//
-//     INSTANA_SECRETS := <matcher>:<secret>[,<secret>]
-//
-// Where `matcher` is one of:
-// * `equals` - matches a string if it's contained in the secrets list
-// * `equals-ignore-case` is a case-insensitive version of `equals`
-// * `contains` matches a string if it contains any of the secrets list values
-// * `contains-ignore-case` is a case-insensitive version of `contains`
-// * `regex` matches a string if it fully matches any of the regular expressions provided in the secrets list
-//
-// This function returns DefaultSecretsMatcher() if there is no matcher configuration provided.
-func parseInstanaSecrets(matcherStr string) (Matcher, error) {
-	if matcherStr == "" {
-		return DefaultSecretsMatcher(), nil
-	}
-
-	ind := strings.Index(matcherStr, ":")
-	if ind < 0 {
-		return nil, fmt.Errorf("malformed INSTANA_SECRETS configuration: %q", matcherStr)
-	}
-
-	matcher, config := strings.TrimSpace(matcherStr[:ind]), strings.Split(matcherStr[ind+1:], ",")
-
-	return NamedMatcher(matcher, config)
-}
-
-// parseInstanaExtraHTTPHeaders parses the tags string passed via INSTANA_EXTRA_HTTP_HEADERS.
-// The header names are expected to come in a semicolon-separated list:
-//
-//     INSTANA_EXTRA_HTTP_HEADERS := header1[;header2;...]
-//
-// Any leading and trailing whitespace characters will be trimmed from header names.
-func parseInstanaExtraHTTPHeaders(headersStr string) []string {
-	var headers []string
-	for _, h := range strings.Split(headersStr, ";") {
-		h = strings.TrimSpace(h)
-		if h == "" {
-			continue
-		}
-
-		headers = append(headers, h)
-	}
-
-	return headers
-}
