+++ util_internal_test.go
@@ -4,10 +4,8 @@ import (
 	"errors"
 	"io/ioutil"
 	"os"
-	"regexp"
 	"testing"
 
-	"github.com/instana/go-sensor/secrets"
 	"github.com/instana/testify/assert"
 	"github.com/instana/testify/require"
 )
@@ -192,89 +190,5 @@ eth0	00000000	010011AC	0003	0	0	0	00000000	0	0	0
 			require.NoError(t, err)
 			assert.Equal(t, test.expected, gateway)
 		}()
-	}
-}
-
-func TestParseInstanaTags(t *testing.T) {
-	examples := map[string]struct {
-		Value    string
-		Expected map[string]interface{}
-	}{
-		"empty":                   {"", nil},
-		"single tag, empty key":   {"=value", nil},
-		"single tag, no value":    {"key", map[string]interface{}{"key": nil}},
-		"single tag, empty value": {"key=", map[string]interface{}{"key": ""}},
-		"single tag, with value":  {"key=value", map[string]interface{}{"key": "value"}},
-		"multiple tags, mixed": {
-			`key1,  key2=  , key3   ="",key4=42`,
-			map[string]interface{}{
-				"key1": nil,
-				"key2": "  ",
-				"key3": `""`,
-				"key4": "42",
-			},
-		},
-	}
-
-	for name, example := range examples {
-		t.Run(name, func(t *testing.T) {
-			assert.Equal(t, example.Expected, parseInstanaTags(example.Value))
-		})
-	}
-}
-
-func TestParseInstanaSecrets(t *testing.T) {
-	regexMatcher, err := secrets.NewRegexpMatcher(regexp.MustCompile("a|b|c"), regexp.MustCompile("d"))
-	require.NoError(t, err)
-
-	examples := map[string]struct {
-		Value    string
-		Expected Matcher
-	}{
-		"empty":                {"", DefaultSecretsMatcher()},
-		"equals":               {"equals:a,b,c", secrets.NewEqualsMatcher("a", "b", "c")},
-		"equals-ignore-case":   {"equals-ignore-case:a,b,c", secrets.NewEqualsIgnoreCaseMatcher("a", "b", "c")},
-		"contains":             {"contains:a,b,c", secrets.NewContainsMatcher("a", "b", "c")},
-		"contains-ignore-case": {"contains-ignore-case:a,b,c", secrets.NewContainsIgnoreCaseMatcher("a", "b", "c")},
-		"regexp":               {"regex:a|b|c,d", regexMatcher},
-	}
-
-	for name, example := range examples {
-		t.Run(name, func(t *testing.T) {
-			m, err := parseInstanaSecrets(example.Value)
-			require.NoError(t, err)
-			assert.Equal(t, example.Expected, m)
-		})
-	}
-}
-
-func TestParseInstanaSecrets_Error(t *testing.T) {
-	examples := map[string]string{
-		"unknown matcher": "magic:pew,pew",
-		"malformed":       "equals;a,b,c",
-	}
-
-	for name, example := range examples {
-		t.Run(name, func(t *testing.T) {
-			_, err := parseInstanaSecrets(example)
-			assert.Error(t, err)
-		})
-	}
-}
-
-func TestParseInstanaExtraHTTPHeaders(t *testing.T) {
-	examples := map[string]struct {
-		Value    string
-		Expected []string
-	}{
-		"empty":    {"", nil},
-		"one":      {"a", []string{"a"}},
-		"multiple": {"a; ;  b  ;c;", []string{"a", "b", "c"}},
-	}
-
-	for name, example := range examples {
-		t.Run(name, func(t *testing.T) {
-			assert.Equal(t, example.Expected, parseInstanaExtraHTTPHeaders(example.Value))
-		})
 	}
 }
